Uint to float

+++ Orbit/Assets/Scripts/Entities/Interfaces/IMovingEntity.cs
@@ -2,6 +2,6 @@
 {
     public interface IMovingEntity
     {
-        uint Speed { get; }
+        float Speed { get; }
     }
 }
